Refactor skymap_objects.py into modular package
